phase1/requirements.txt
# Phase 1: Data Ingestion Foundation
# Python dependencies

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.12.1

# RSS Feed Parsing
feedparser==6.0.10

# Web Scraping
scrapy==2.11.0
play-scraper==1.2.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2

# Caching
redis==5.0.1

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Task Queue
celery==5.3.4

# Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Utilities
python-dateutil==2.8.2
tqdm==4.66.1
